+++ src/services/Emails.php
@@ -341,6 +341,7 @@ class Emails extends Component
             $emailRecord = $this->_getEmailRecord($emailUid);
             $isNewEmail = $emailRecord->getIsNewRecord();
             $store = Plugin::getInstance()->getStores()->getStoreByUid($data['store']);
+            $renderSite = array_key_exists('renderSite', $data) && $data['renderSite'] !== null ? Craft::$app->getSites()->getSiteByUid($data['renderSite']) : null;
 
             $emailRecord->storeId = $store->id;
             $emailRecord->name = $data['name'];
@@ -358,6 +359,7 @@ class Emails extends Component
             $emailRecord->uid = $emailUid;
             $emailRecord->pdfId = $pdfUid ? Db::idByUid(Table::PDFS, $pdfUid) : null;
             $emailRecord->language = $data['language'] ?? EmailRecord::LOCALE_ORDER_LANGUAGE;
+            $emailRecord->renderSiteId = $renderSite?->id ?? null;
 
             $emailRecord->save(false);
 
@@ -477,6 +479,7 @@ class Emails extends Component
         $originalLanguage = Craft::$app->language;
         $originalFormattingLanguage = Craft::$app->formattingLocale;
         $emailLanguage = $email->getRenderLanguage($order);
+        $emailSite = $email->getRenderSite($order);
 
         Locale::switchAppLanguage($emailLanguage);
 
@@ -783,6 +786,9 @@ class Emails extends Component
             }
         }
 
+        $originalSiteId = Craft::$app->getSites()->getCurrentSite()->id;
+        Craft::$app->getSites()->setCurrentSite($emailSite);
+
         // Render HTML body
         try {
             $body = $view->renderTemplate($templatePath, $renderVariables);
@@ -799,6 +805,7 @@ class Emails extends Component
             ]);
             Craft::error($error, __METHOD__);
 
+            Craft::$app->getSites()->setCurrentSite($originalSiteId);
             Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
             $view->setTemplateMode($oldTemplateMode);
             $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -823,6 +830,7 @@ class Emails extends Component
                 ]);
                 Craft::error($error, __METHOD__);
 
+                Craft::$app->getSites()->setCurrentSite($originalSiteId);
                 Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
                 $view->setTemplateMode($oldTemplateMode);
                 $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -850,6 +858,7 @@ class Emails extends Component
 
                 Craft::info($notice, __METHOD__);
 
+                Craft::$app->getSites()->setCurrentSite($originalSiteId);
                 Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
                 $view->setTemplateMode($oldTemplateMode);
                 $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -869,6 +878,7 @@ class Emails extends Component
 
                 Craft::error($error, __METHOD__);
 
+                Craft::$app->getSites()->setCurrentSite($originalSiteId);
                 Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
                 $view->setTemplateMode($oldTemplateMode);
                 $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -888,6 +898,7 @@ class Emails extends Component
 
             Craft::error($error, __METHOD__);
 
+            Craft::$app->getSites()->setCurrentSite($originalSiteId);
             Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
             $view->setTemplateMode($oldTemplateMode);
             $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -906,6 +917,7 @@ class Emails extends Component
             ]));
         }
 
+        Craft::$app->getSites()->setCurrentSite($originalSiteId);
         Locale::switchAppLanguage($originalLanguage, $originalFormattingLanguage->id);
         $view->setTemplateMode($oldTemplateMode);
         $generalConfig->generateTransformsBeforePageLoad = $generateTransformsBeforePageLoad;
@@ -957,6 +969,7 @@ class Emails extends Component
                 'emails.pdfId',
                 'emails.plainTextTemplatePath',
                 'emails.recipientType',
+                'emails.renderSiteId',
                 'emails.replyTo',
                 'emails.senderAddress',
                 'emails.senderName',